+++ instrumentation/opentelemetry-instrumentation-pymongo/tests/test_pymongo.py
@@ -14,11 +14,13 @@
 
 from unittest import mock
 
+from opentelemetry import context
 from opentelemetry import trace as trace_api
 from opentelemetry.instrumentation.pymongo import (
     CommandTracer,
     PymongoInstrumentor,
 )
+from opentelemetry.instrumentation.utils import _SUPPRESS_INSTRUMENTATION_KEY
 from opentelemetry.semconv.trace import SpanAttributes
 from opentelemetry.test.test_base import TestBase
 
@@ -89,6 +91,31 @@ class TestPymongo(TestBase):
         self.assertTrue(mock_span.is_recording.called)
         self.assertFalse(mock_span.set_attribute.called)
         self.assertFalse(mock_span.set_status.called)
+
+    def test_suppression_key(self):
+        mock_tracer = mock.Mock()
+        mock_span = mock.Mock()
+        mock_span.is_recording.return_value = True
+        mock_tracer.start_span.return_value = mock_span
+        mock_event = MockEvent({})
+        mock_event.command.get = mock.Mock()
+        mock_event.command.get.return_value = "dummy"
+
+        token = context.attach(
+            context.set_value(_SUPPRESS_INSTRUMENTATION_KEY, True)
+        )
+
+        try:
+            command_tracer = CommandTracer(mock_tracer)
+            command_tracer.started(event=mock_event)
+            command_tracer.succeeded(event=mock_event)
+        finally:
+            context.detach(token)
+
+        # if suppression key is set, CommandTracer methods return immediately, so command.get is not invoked.
+        self.assertFalse(
+            mock_event.command.get.called  # pylint: disable=no-member
+        )
 
     def test_failed(self):
         mock_event = MockEvent({})